Relocate integration tests
